refactor(persistence): relocate wirePersistence to core/persistence/
Move the ~440-LOC `wirePersistence` closure plus its three private
helpers (`isEmptyContainer`, `collectPersistedLeafPaths`,
`isDescendantPathKey`) out of `use-abstract-form.ts` (~526 LOC trimmed)
into a focused `core/persistence/wire-persistence.ts` module — putting
it next to the rest of the persistence layer (storage adapters,
sensitive-name filter, opt-in registry, read/write/orphan-sweep
helpers it composes).

`wirePersistence` was already module-scoped (no closure capture over
`useAbstractForm`'s locals), so the move is a verbatim relocation
modulo the import block. The composable file slims down ~36% and now
imports `wirePersistence` from the same neighborhood as
`normalizePersistConfig`, `PERSISTENCE_MODULE_KEY`,
`resolveStorageKeyBase`, and the orphan-sweep helpers. The relocated
module owns its dependency surface: storage adapter promise, schema
fingerprint hash, debounced writer, sparse payload buildup, opt-in
path snapshotting, all the post-hydration revalidation choreography.

The host file drops unused imports — `toRaw`,
`DEFAULT_PERSISTENCE_DEBOUNCE_MS`, `mergeSparseHydration`,
`buildPersistedPayload`, `cleanupOrphanKeys`, `createDebouncedWriter`,
`filterErrorsByPaths`, `getStorageAdapter`, `pluckPaths`,
`readPersistedPayload`, `stripUnacknowledgedSensitiveLeaves`,
`canonicalizePath` / `Path` / `PathKey`, `deleteAtPath` / `getAtPath` /
`setAtPath` / `isPlainRecord`, `ValidationError`.

The new module routes the `disposed` flag through an `isDisposed()`
function-call indirection so TS's control-flow narrowing can't
collapse the second-and-later post-await `if (isDisposed()) return`
checks to "always falsy" (the strict-rule block at the new path
enforces `@typescript-eslint/no-unnecessary-condition`, which the
composables exemption let slip in the original location).

Behavior-neutral: persistence + blank-persistence suites (54 tests)
and the full 3547-test suite stay green unchanged.
